new: добавлена логика работы с обработкой тестирования (назначение теста и обрабока ответов от пользователя)!wip

+++ src/main/java/com/kamilla/deppplom/tests/model/Test.java
@@ -1,10 +1,7 @@
 package com.kamilla.deppplom.tests.model;
 
 import com.kamilla.deppplom.discipline.Discipline;
-import com.kamilla.deppplom.question.model.Question;
-import com.kamilla.deppplom.question.model.Selection;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
 
 import java.util.List;
@@ -14,7 +11,7 @@ import java.util.List;
 public class Test {
     private int id;
     private String title;
-    private int minimumPoints;
+    private float minimumPoints;
     private List<TestVersion> versions;
     private Discipline discipline;
 }
